feat(minibf): governance dreps endpoint - additional drep data and metadata fetching
